Update Dockerfile adding back in `public` directory
See #8516

In the build I mistakenly prevented copying in the `public` directory
thinking that it was generated during the build process.

Actually it held our static assets and needed to be there. Images were
missing, for example.

The same Dockerfile has been updated again but this time I have not
prevented Docker from incorporating the `public` directory